requirements.txt
numpy>=1.24
scipy>=1.10
matplotlib>=3.8
PyQt5>=5.15
pyqtgraph>=0.13
pyrtlsdr>=0.3.0
python-osc>=1.8
sounddevice>=0.4
scikit-learn>=1.3

# S.A.R geospatial (Phase 1+)
shapely>=2.0
pyproj>=3.6
requests>=2.28
Pillow>=9.0

# OSC sonification (optional — system works without it)
python-osc>=1.8
